pyproject.toml
[project]
name = "f1-digital-twin-monza"
version = "0.2.0"
description = "Spatial telemetry reconstruction for F1 qualifying sessions"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["f1", "telemetry", "digital-twin", "motorsport", "data-engineering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "fastf1>=3.1.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "ruff>=0.1.0"]

[project.scripts]
f1twin = "main:main"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
log_cli = true
log_cli_level = "WARNING"
